pyproject.toml
[project]
name = "firware-based-architecture-for-slec"
version = "0.1.0"
description = "WiP implementation of the ISIE 2023 paper"
authors = [
    {name = "Max Thoma"}
]
license = {text = "BSD-3-Clause"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib (>=3.10.5,<4.0.0)",
    "mosaik (>=3.5.0,<4.0.0)",
    "paho-mqtt (>=1.0.0,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "pvlib (>=0.13.0,<0.14.0)",
    "mosaik-csv (>=1.0.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "mosaik-api (>=3.0.6,<4.0.0)",
    "setuptools (>=80.9.0,<81.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
